chore: move WaitServer
