+++ thrift_streaming/thrift_slim_others_test.go
@@ -1,5 +1,5 @@
-//go:build !amd64 && windows && !go1.16 && go1.22 && disablefrugal
-// +build !amd64,windows,!go1.16,go1.22,disablefrugal
+//go:build !amd64 || windows || !go1.16 || go1.22 || disablefrugal
+// +build !amd64 windows !go1.16 go1.22 disablefrugal
 
 // Copyright 2023 CloudWeGo Authors
 //
